% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpProblems.R
\name{booleanGrammar}
\alias{booleanGrammar}
\title{A constant function with a boolean grammar.}
\usage{
booleanGrammar()
}
\value{
A named list with $filename and  $BNF,
        the grammar of a boolean grammar with two variables.
}
\description{
For the distribution of examples of BNF in grammars.
}
\details{
Imported from package xegaBNF for use in examples.
}
\examples{
booleanGrammar()
}
\seealso{
Other Grammar: 
\code{\link{compileBNF}()}
}
\concept{Grammar}
