% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaPermProblems.R
\docType{data}
\name{lau15}
\alias{lau15}
\title{The problem environment lau15}
\format{
An object of class \code{list} of length 14.
}
\usage{
lau15
}
\description{
15 abstract cities for which a traveling salesman solution is sought.
Solution: A path with a length of 291.
}
\examples{
names(lau15)
lau15$genelength()
}
\references{
Lau, H. T. (1986):
\emph{Combinatorial Heuristic Algorithms in FORTRAN}.
Springer, 1986.
<doi:10.1007/978-3-642-61649-5>
}
\seealso{
Other Problem Environment: 
\code{\link{NewEnvXOR}()},
\code{\link{Parabola2D}},
\code{\link{Parabola2DEarly}}
}
\concept{Problem Environment}
\keyword{datasets}
