% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgaProblems.R
\docType{data}
\name{Parabola2DEarly}
\alias{Parabola2DEarly}
\title{Problem environment for a 2-dimensional quadratic parabola.}
\format{
An object of class \code{list} of length 9.
}
\usage{
Parabola2DEarly
}
\value{
A problem environment (see \link{Parabola2D}).
        \code{Parabola2DEarly$terminate(solution, lF)} 
        is a test function which returns true if the \code{solution} 
        is in an epsilon environment of a known solution. 
        To invoke this function, use \code{xegaRun( ..., early=TRUE, ...)}.
        The epsilon which determines 
        the length of the interval as a fraction
        of the known optimal solution is configured by  
        e.g. \code{xegaRun( ..., terminationEps=0.001, ...)}.
}
\description{
An example of a problem environment with an 
             early termination condition.
}
\seealso{
Other Problem Environment: 
\code{\link{NewEnvXOR}()},
\code{\link{Parabola2D}},
\code{\link{lau15}}
}
\concept{Problem Environment}
\keyword{datasets}
