% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgXDecodeFactory.R
\name{sgXDecodeGeneFactory}
\alias{sgXDecodeGeneFactory}
\title{Factory for configuring a gene-dependent DecodeGene function.}
\usage{
sgXDecodeGeneFactory(algorithm = "sga", method = "DecodeGene")
}
\arguments{
\item{algorithm}{"sga", "sgde", "sgperm", "sge", "sgede", 
"sgp". Default: "sga".}

\item{method}{Method. Default: "DecodeGene".}
}
\value{
Decode function for the selected algorithm from the correct package.
}
\description{
A gene-specific decoder must be provided.
}
\examples{
sgXDecodeGeneFactory(algorithm="sgperm", method="DecodeGene")

}
\seealso{
Other Configuration: 
\code{\link{sgXCrossoverFactory}()},
\code{\link{sgXGeneMapFactory}()},
\code{\link{sgXInitGeneFactory}()},
\code{\link{sgXMutationFactory}()},
\code{\link{sgXReplicationFactory}()}
}
\concept{Configuration}
