% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgXInitFactory.R
\name{sgXInitGeneFactory}
\alias{sgXInitGeneFactory}
\title{Factory for configuring a gene-dependent InitGene function.}
\usage{
sgXInitGeneFactory(algorithm = "sga", method = "InitGene")
}
\arguments{
\item{algorithm}{Algorithm. 
Available: "sga", "sgde", "sgperm", "sge", sgp". 
Default: "sga".}

\item{method}{Method. Default: "InitGene".
For sgp, method = "InitGene" or "InitGeneGe".}
}
\value{
InitGene function from the correct package.
}
\description{
Factory for configuring a gene-dependent InitGene function.
}
\examples{
sgXInitGeneFactory(algorithm="sgperm")

}
\seealso{
Other Configuration: 
\code{\link{sgXCrossoverFactory}()},
\code{\link{sgXDecodeGeneFactory}()},
\code{\link{sgXGeneMapFactory}()},
\code{\link{sgXMutationFactory}()},
\code{\link{sgXReplicationFactory}()}
}
\concept{Configuration}
