% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaResult.R
\name{xegaAnyTimeResult}
\alias{xegaAnyTimeResult}
\title{Writes xega results after each iteration to a rds file.}
\usage{
xegaAnyTimeResult(
  mlT,
  pp,
  ft,
  lF,
  allsolutions,
  popStat,
  evalFail,
  GAconfiguration,
  path
)
}
\arguments{
\item{mlT}{Main loop timer function.}

\item{pp}{Population.}

\item{ft}{Fitness vector.}

\item{lF}{Local function configuration.}

\item{allsolutions}{Boolean.}

\item{popStat}{Population statistics
(mean, min, Q1, median, Q3, max, var, mad).}

\item{evalFail}{Number of evaluation failures.}

\item{GAconfiguration}{Configuration of the genetic algorithm.}

\item{path}{File path.}
}
\value{
Invisible 0.
}
\description{
Writes xega results after each iteration to a rds file.
}
\details{
The file \code{xegaAnyTimeResult.rds} is overwritten 
         after each generation.
         The function is intended to be used with the defaults
         (\code{xegaAnyTimeResult()}).
}
