% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaVersion.R
\name{xegaVersion}
\alias{xegaVersion}
\title{About this version.}
\usage{
xegaVersion(verbose = TRUE)
}
\arguments{
\item{verbose}{Boolean. If \code{TRUE} (Default), print 
package information and version number to console.}
}
\value{
Version number (invisible).
}
\description{
About this version.
}
\examples{
xegaVersion()
}
